+++ README.md
@@ -6,3 +6,4 @@ File structure:
 - `.gitignore`: Defines the files to be excluded from the git repo. This includes `.pyc` files and other IDE/editor-specific config files. Modify this file if your IDE/editor generates irrelevant files.
 - `src/`: Contains the source files.
 - `res/`: Contains sounds, images, maps, etc.
+- `doc/`: Contains documentations.
