--- levelData/level1.txt
@@ -1,11 +1,0 @@
-8
-13
-1234012340123
-1234012340123
-1234012340123
-1234012340123
-1234012340123
-1234012340123
-1234012340123
-1234012340123
-
